pyproject.toml
[project]
name = "idl-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "scipy>=1.11.0",
    "scikit-image>=0.22.0",
    "matplotlib>=3.10.8",
    "tqdm>=4.65.0",
    "astropy>=6.0.0",
    "torch>=2.11.0",
    "datasets>=2.19.1",
    "opencv-python>=4.13.0.92",
    "torchvision>=0.26.0",
    "streamlit>=1.56.0",
]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu128"]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
